pyproject.toml
[project]
name = "english-skill-app"
version = "0.1.0"
description = "英会話能力測定AIアプリ"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flet[all]==0.25.0",
    "openai>=1.12.0",
    "azure-cognitiveservices-speech>=1.38.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pyinstaller>=6.3.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "pydub>=0.25.1",
    "audioop-lts>=0.2.2",  # Python 3.13+でaudioopの代替として必要
    "aiofiles>=23.2.0",  # 非同期ファイルI/O
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "types-aiofiles>=24.1.0",
    "pydub-stubs>=0.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pydub-stubs>=0.25.1.6",
    "types-aiofiles>=25.1.0.20251011",
]

[tool.uv.sources]

